ext: adding notes
