+++ src/configs/config.yaml
@@ -6,6 +6,7 @@ project: ENSOTrend-beta
 user: sdat2
 animate: true # whether to make animations after run.
 wandb: true # whether or not to sync to weights and biases.
+notes: NA
 # name: k_days_${atm.k_days} # name needs to be unique otherwise will overwrite.
 name: k_days_${atm.k_days}_eps_days_${atm.eps_days}_e_frac_${atm.e_frac}_c_d_${coup.c_d}_vary_cloud_const_${atm.vary_cloud_const}_${atm.mem}
 # name: eps_days_${atm.eps_days}
